+++ suager/utils/tbl.py
@@ -37,7 +37,8 @@ def get_clan(guild, db):
             "temple_levels": json.dumps([1] * len(tbl_data.totems)),
             "nuts": 0,
             "coins": 0,
-            "upgrade_points": 0
+            "upgrade_points": 0,
+            "usage": False
         }, False
     return clan, True
 
@@ -47,11 +48,15 @@ async def tbl_game(ctx, db):
     now_dt = time.now(None)
     player, wp = get_player(ctx.author, db, now_ts)
     clan, wc = get_clan(ctx.guild, db)
+    if wc:
+        if clan["usage"]:
+            return await general.send("It seems that TBL is currently running in this server, please try again later.", ctx.channel)
+        db.execute("UPDATE tbl_clan SET usage=1 WHERE gid=?", (ctx.guild.id,))
     try:
         lid = player["location"]
         level = player['level']
         loc = get_location(lid, level)
-        stuff = f"{time.time(None, _tz=True)} > {ctx.author.name} > TBL Initiated"
+        stuff = f"{ctx.author.name} > TBL Initiated"
         energy, regen_time = player["energy"], player["time"]
         energy, regen_time = regen_energy(energy, regen_time, level, int(now_ts))
         if energy < loc["energy"]:
@@ -62,7 +67,15 @@ async def tbl_game(ctx, db):
         new_level = player["level"]
         _runs = 0
         _act = get_activity(loc['activity'])
-        while energy >= loc["energy"] and _runs <= 100:
+        act = random.randint(int(_act * 0.9), int(_act * 1.1))
+        if act < 100:
+            people = 15
+        elif 100 <= act < 1000:
+            people = int(act / random.uniform(3, 5))
+        else:
+            people = int(act / random.uniform(7, 20))
+        people = 1500 if people > 1500 else people
+        while energy >= loc["energy"] and _runs < 250:
             runs += 1
             _runs += 1
             cool = random.random() > 0.84
@@ -70,13 +83,7 @@ async def tbl_game(ctx, db):
             life = random.random()
             life_req = loc['dr'] / (1 + (new_level - 1) / 64)
             live = life > life_req
-            act = random.randint(int(_act * 0.9), int(_act * 1.1))
-            if act < 100:
-                people = 15
-            elif 100 <= act < 1000:
-                people = int(act / random.uniform(3, 5))
-            else:
-                people = int(act / random.uniform(7, 20))
+            people = int(people * random.uniform(0.9, 1.1))
             place = 0
             saves = 0
             nuts, xp, points, sh_xp = 0, 0, 0, 0
@@ -108,9 +115,9 @@ async def tbl_game(ctx, db):
                 temple = temples[i]
                 if temple["expiry"] > now_ts or i == 0:
                     if temple["id"] == 1:
-                        nuts *= 1.08 + 0.02 * temple_levels[0]
+                        nuts *= 1.06 + 0.04 * temple_levels[0]
                     elif temple["id"] == 2:
-                        xp *= 1.06 + 0.04 * temple_levels[1]
+                        xp *= 1.07 + 0.03 * temple_levels[1]
                     elif temple["id"] == 3:
                         sh_xp *= 1.075 + 0.025 * temple_levels[2]
             _lvl_len_int = loc['ll']
@@ -137,7 +144,7 @@ async def tbl_game(ctx, db):
             place_str = f"You came **#{place:,}/{people - 1:,}**"
             saves_str = f"Saves: **{saves:,}**"
             sh_xp_str = f"\nShaman XP: **{sh_xp:,.0f}**" if cool else ""
-            message_data = f"{time.time()} > {ctx.author.name} > TBL\nRound: {_runs}\nLocation: **{loc['name']}**\n" \
+            message_data = f"**{ctx.author.name}** > TBL\nRound: **{_runs}**\nLocation: **{loc['name']}**\n" \
                            f"Energy remaining: **{energy:,.0f}**\n\nThis round's results:\nLevel length: **{lvl_len}**\n"
             if live:
                 if not cool:
@@ -151,7 +158,7 @@ async def tbl_game(ctx, db):
             message_data += f"\nRewards:\nNuts: **{nuts:,.0f}** - New total: **{player['nuts']:,}**\nXP: **{xp:,.0f}** - New total: **{player['xp']:,}** - " \
                             f"Level **{new_level}**{sh_xp_str}"
             await message.edit(content=message_data)
-            await asyncio.sleep(2)
+            await asyncio.sleep(1.5)
         new_level, ld, title = xp_level(level, player['xp'])
         level_up = f"__You are now Level **{new_level}**! New title: **{title}**__\n" if ld > 0 else ""
         new_sh_level = sh_level(player['sh_xp'])
@@ -185,6 +192,8 @@ async def tbl_game(ctx, db):
         if ctx.channel.id == 610482988123422750:
             await general.send(general.traceback_maker(e), ctx.channel)
         return await general.send(f"An error has occurred.\n`{type(e).__name__}: {e}`\nData was reverted.", ctx.channel)
+    finally:
+        db.execute("UPDATE tbl_clan SET usage=0 WHERE gid=?", (ctx.guild.id,))
 
 
 def get_location(lid: int, level: int):
@@ -256,3 +265,14 @@ def clan_level(old_lvl: int, xp: int):
         else:
             break
     return new_lvl, new_lvl - old_lvl
+
+
+def get_season():
+    now = time.now()
+    for season, data in tbl_data.seasons.items():
+        start, end = data
+        if end < now:
+            continue
+        elif start > now:
+            continue
+        return season